+++ packages/navigation/src/NavigationItem.tsx
@@ -1,10 +1,18 @@
 import type { NavItemProps } from './NavItem';
 import type { NavItemWithTooltipProps } from './NavItemWithTooltip';
-import type { ReactElement } from 'react';
+import type { ReactElement, MouseEvent } from 'react';
 
-type NavigationItemProps = NavItemProps &
-  NavItemWithTooltipProps & {
+// The public interface for `NavigationItem` should include an `onClick` prop
+// that receives an additional `state` param (determined by either
+// `NavigationList` or `NavigationMenuDropdown`). The underlying `NavItem` and
+// `NavItemWithTooltip` implementations pass their `onClick` directly through
+// to `NavLink`, so we need to omit their `onClick` implementations for type
+// correctness.
+type NavigationItemProps = Omit<NavItemProps, 'onClick'> &
+  Omit<NavItemWithTooltipProps, 'onClick'> & {
     tooltip?: boolean | ReactElement;
+  } & {
+    onClick?(e: MouseEvent, state: { collapsed: boolean }): void;
   };
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars