+++ packages/navigation/src/index.ts
@@ -1,6 +1,13 @@
-export type { NavigationProps } from './Navigation';
+export type { NavigationProps, NavigationState } from './Navigation';
 export type { NavigationItemProps } from './NavigationItem';
 /* c8 ignore start */
 export { Navigation } from './Navigation';
 export { NavigationItem } from './NavigationItem';
+/* c8 ignore stop */
+
+export type { NavigationProps as LegacyNavigationProps } from './legacy/Navigation';
+export type { NavigationItemProps as LegacyNavigationItemProps } from './legacy/NavigationItem';
+/* c8 ignore start */
+export { Navigation as LegacyNavigation } from './legacy/Navigation';
+export { NavigationItem as LegacyNavigationItem } from './legacy/NavigationItem';
 /* c8 ignore stop */